LPS-55928 move query visitor to query package
