+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/internal/query/BooleanQueryTranslatorImpl.java
@@ -18,7 +18,7 @@ import com.liferay.portal.kernel.search.BooleanClause;
 import com.liferay.portal.kernel.search.BooleanClauseOccur;
 import com.liferay.portal.kernel.search.BooleanQuery;
 import com.liferay.portal.kernel.search.Query;
-import com.liferay.portal.kernel.search.QueryVisitor;
+import com.liferay.portal.kernel.search.query.QueryVisitor;
 import com.liferay.portal.search.elasticsearch.query.BooleanQueryTranslator;
 
 import org.elasticsearch.index.query.BoolQueryBuilder;
