+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/internal/query/ElasticsearchQueryTranslator.java
@@ -16,12 +16,12 @@ package com.liferay.portal.search.elasticsearch.internal.query;
 
 import com.liferay.portal.kernel.search.BooleanQuery;
 import com.liferay.portal.kernel.search.Query;
-import com.liferay.portal.kernel.search.QueryVisitor;
 import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.TermQuery;
 import com.liferay.portal.kernel.search.TermRangeQuery;
 import com.liferay.portal.kernel.search.WildcardQuery;
 import com.liferay.portal.kernel.search.query.QueryTranslator;
+import com.liferay.portal.kernel.search.query.QueryVisitor;
 import com.liferay.portal.search.elasticsearch.query.BooleanQueryTranslator;
 import com.liferay.portal.search.elasticsearch.query.TermQueryTranslator;
 import com.liferay.portal.search.elasticsearch.query.TermRangeQueryTranslator;
