+++ portal-service/src/com/liferay/portal/kernel/search/BaseQueryImpl.java
@@ -14,6 +14,8 @@
 
 package com.liferay.portal.kernel.search;
 
+import com.liferay.portal.kernel.search.query.QueryVisitor;
+
 /**
  * @author Michael C. Han
  */
